requirements.txt
img2dataset==1.28.0
img2vec_pytorch==1.0.1
matplotlib==3.2.2
numpy==1.21.5
pandas==1.3.5
Pillow==9.1.0
torchvision==0.11.1+cu111
tqdm==4.63.0
requests
bs4
pathlib
gazpacho
alive-progress
click
pylint
ez_setup
black